requirements.txt
streamlit==1.29.0
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
pandas==2.1.4
openpyxl==3.1.2
google-auth==2.25.2
